Clear the item info cache on every BAG_UPDATE
Most of the item information is variable, such as count, quality and more, and
can change in an item's lifespan of a slot.
We cached everything for speed, but the cache needs to be updated, so we clear
it whenever there's a possibility for a change.

Still want to take a further look into C_Item API.