Base itemMixin more on ItemMixin and C_Item
We can't use ItemMixin directly because it assumes to much on the cached data
(see ItemMixin:GetStaticBackingItem), which can be volatile in a bag scenario.

Not the easiest to evaluate the performance difference, but C_Item and ItemMixin
is easier to deal with when it comes to loading items.

Also added a load of helpful comments on the load-necessity and data volatility.